T3-function/requirements.txt
azure-functions
psycopg2-binary==2.8.4
